fix: uniform countdown cadence on widget and settings
The displayed seconds were sampled by a 1s interval whose phase drifts
against the countdown's real second boundaries (and reset on every
server sync/donation), so a value could hold ~2s then jump ahead.
Seconds are now derived from endTime by a shared hook that samples at
100ms and re-renders only when the integer changes, putting every
transition within ~100ms of the true boundary.
